move DownedBossSystem into own folder
